# api/src/generated/openapi.yaml
openapi: 3.0.0
components:
...
  schemas:
...
    RSVP:
      description: RSVP model
      properties:
        userId:
          type: string
          description: User ID who made the RSVP
        status:
          type: string
          enum:
            - going
            - maybe
            - notGoing
          description: RSVP status
        memo:
          type: string
          description: Optional memo or note
        createdAt:
          type: string
          format: date-time
          description: Timestamp when created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when updated
      required:
        - userId
        - status
        - memo
        - createdAt
        - updatedAt
      type: object
      additionalProperties: false
    Comment:
      description: Comment model
      properties:
        id:
          type: string
          description: Unique comment identifier
        userId:
          type: string
          description: ID of the user who made the comment
        text:
          type: string
          description: Comment text
        photos:
          items:
            type: string
          type: array
          description: Array of photo URLs attached to comment
        createdAt:
          type: string
          format: date-time
          description: Timestamp when created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when updated
      required:
        - id
        - userId
        - text
        - photos
        - createdAt
        - updatedAt
      type: object
      additionalProperties: false
    EventDetailed:
      description: Event with RSVPs and comments (detailed view)
      properties:
        id:
          type: string
          description: Unique event identifier
        groupId:
          type: string
          description: ID of the group this event belongs to
        createdBy:
          type: string
          description: ID of the user who created this event
        updatedBy:
          type: string
          description: ID of the user who last updated this event
        title:
          type: string
          description: Event title
        subtitle:
          type: string
          nullable: true
          description: Event subtitle
        description:
          type: string
          nullable: true
          description: Event description
        coverPhotos:
          items:
            type: string
          type: array
          description: Array of cover photo URLs
        start:
          type: string
          format: date-time
          description: Event start date/time
        end:
          type: string
          format: date-time
          description: Event end date/time
        isAllDay:
          type: boolean
          nullable: true
          description: Whether this is an all-day event
        location:
          type: string
          nullable: true
          description: Event location
        minAttendees:
          type: number
          format: double
          nullable: true
          description: Minimum number of attendees required
        deadline:
          type: string
          format: date-time
          nullable: true
          description: RSVP deadline
        allowMaybe:
          type: boolean
          description: Whether 'maybe' RSVPs are allowed
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the event was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the event was last updated
        rsvps:
          items:
            $ref: "#/components/schemas/RSVP"
          type: array
          description: Array of RSVPs for this event
        comments:
          items:
            $ref: "#/components/schemas/Comment"
          type: array
          description: Array of comments on this event
      required:
        - id
        - groupId
        - createdBy
        - updatedBy
        - title
        - coverPhotos
        - start
        - end
        - allowMaybe
        - createdAt
        - updatedAt
        - rsvps
        - comments
      type: object
      additionalProperties: false
    Event:
      description: Event model - represents a scheduled event
      properties:
        id:
          type: string
          description: Unique event identifier
        groupId:
          type: string
          description: ID of the group this event belongs to
        createdBy:
          type: string
          description: ID of the user who created this event
        updatedBy:
          type: string
          description: ID of the user who last updated this event
        title:
          type: string
          description: Event title
        subtitle:
          type: string
          nullable: true
          description: Event subtitle
        description:
          type: string
          nullable: true
          description: Event description
        coverPhotos:
          items:
            type: string
          type: array
          description: Array of cover photo URLs
        start:
          type: string
          format: date-time
          description: Event start date/time
        end:
          type: string
          format: date-time
          description: Event end date/time
        isAllDay:
          type: boolean
          nullable: true
          description: Whether this is an all-day event
        location:
          type: string
          nullable: true
          description: Event location
        minAttendees:
          type: number
          format: double
          nullable: true
          description: Minimum number of attendees required
        deadline:
          type: string
          format: date-time
          nullable: true
          description: RSVP deadline
        allowMaybe:
          type: boolean
          description: Whether 'maybe' RSVPs are allowed
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the event was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the event was last updated
      required:
        - id
        - groupId
        - createdBy
        - updatedBy
        - title
        - coverPhotos
        - start
        - end
        - allowMaybe
        - createdAt
        - updatedAt
      type: object
      additionalProperties: false
    EventInput:
      description: Input for creating a new event
      properties:
        id:
          type: string
        groupId:
          type: string
        createdBy:
          type: string
        title:
          type: string
        subtitle:
          type: string
        description:
          type: string
        coverPhotos:
          items:
            type: string
          type: array
        start:
          anyOf:
            - type: string
              format: date-time
            - type: string
        end:
          anyOf:
            - type: string
              format: date-time
            - type: string
        isAllDay:
          type: boolean
        location:
          type: string
        minAttendees:
          type: number
          format: double
        deadline:
          anyOf:
            - type: string
              format: date-time
            - type: string
        allowMaybe:
          type: boolean
      required:
        - id
        - groupId
        - createdBy
        - title
        - start
        - end
      type: object
      additionalProperties: false
    EventUpdate:
      description: Input for updating an event
      properties:
        title:
          type: string
        subtitle:
          type: string
        description:
          type: string
        coverPhotos:
          items:
            type: string
          type: array
        start:
          anyOf:
            - type: string
              format: date-time
            - type: string
        end:
          anyOf:
            - type: string
              format: date-time
            - type: string
        isAllDay:
          type: boolean
        location:
          type: string
        minAttendees:
          type: number
          format: double
        deadline:
          anyOf:
            - type: string
              format: date-time
            - type: string
        allowMaybe:
          type: boolean
        updatedBy:
          type: string
      required:
        - updatedBy
      type: object
      additionalProperties: false
    RSVPInput:
      description: Input for creating/updating an RSVP
      properties:
        userId:
          type: string
        status:
          type: string
          enum:
            - going
            - maybe
            - notGoing
        memo:
          type: string
      required:
        - userId
        - status
      type: object
      additionalProperties: false
    CommentInput:
      description: Input for creating a comment
      properties:
        id:
          type: string
        userId:
          type: string
        text:
          type: string
        photos:
          items:
            type: string
          type: array
      required:
        - id
        - userId
      type: object
      additionalProperties: false
  securitySchemes: {}
info:
  title: BoltUp API
  version: 1.0.0
  description: REST API for BoltUp - A social event planning platform for groups
  license:
    name: ISC
  contact:
    name: BoltUp Team
paths:
...
  /events:
    get:
      operationId: GetEvents
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/EventDetailed"
                type: array
      description: Get all events
      summary: Retrieves events scoped by user's group membership. userId required.
      tags:
        - Events
      security: []
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: groupId
          required: false
          schema:
            type: string
        - in: query
          name: startAfter
          required: false
          schema:
            type: string
        - in: query
          name: startBefore
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            format: double
            type: number
    post:
      operationId: CreateEvent
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
      description: Create a new event
      summary: Creates a new event in a group
      tags:
        - Events
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
  /events/{id}:
    get:
      operationId: GetEvent
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetailed"
      description: Get event by ID
      summary: Retrieves a single event. userId required to verify access.
      tags:
        - Events
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: userId
          required: true
          schema:
            type: string
    put:
      operationId: UpdateEvent
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
      description: Update an event
      summary: Updates an existing event's information
      tags:
        - Events
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
    delete:
      operationId: DeleteEvent
      responses:
        "204":
          description: No Content
      description: Delete an event
      summary: Deletes an event and all associated data
      tags:
        - Events
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
  /events/{id}/rsvps:
    post:
      operationId: UpsertRSVP
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RSVP"
      description: Create or update RSVP
      summary: Creates or updates a user's RSVP for an event
      tags:
        - Events
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RSVPInput"
  /events/{id}/rsvps/{userId}:
    delete:
      operationId: DeleteRSVP
      responses:
        "204":
          description: No Content
      description: Delete RSVP
      summary: Removes a user's RSVP from an event
      tags:
        - Events
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
  /events/{id}/comments:
    get:
      operationId: GetComments
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Comment"
                type: array
      description: Get event comments
      summary: Retrieves all comments for an event
      tags:
        - Events
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
    post:
      operationId: CreateComment
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
      description: Create a comment
      summary: Adds a new comment to an event
      tags:
        - Events
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
  /comments/{id}:
    delete:
      operationId: DeleteComment
      responses:
        "204":
          description: No Content
      description: Delete a comment
      summary: Deletes a comment from an event
      tags:
        - Comments
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
servers:
  - url: http://localhost:3000/api